Add files via upload
M5Stack displays temperature and humidity data from a DHT22 in graphical form and stores the data on an SD card. The data is time stamped using the the time from an NTP server. This also means that a WiFi connection is required. If no WiFi is available the data is stored without a time stamp. The sample rate in my application is 1 reading/min but can be easily changed. It's for use in my house so the temperature range displayed is between 10-30C.